edit and delete methods are added to the articles table

+++ src/app/components/common/table/tableBody.jsx
@@ -1,39 +1,44 @@
 import React from "react";
 import PropTypes from "prop-types";
+import { useSelector } from "react-redux";
+import { getUserById } from "../../../store/users";
+import { getCategoryById } from "../../../store/categories";
 
 const TableBody = ({ article, onEdit, onDelete }) => {
-    console.log("TableBody:", article);
+    const author = useSelector(getUserById(article.author));
+    const { firstName, lastName } = author;
+    const category = useSelector(getCategoryById(article.category));
+    // console.log("TableBody:", category);
     const handleEdit = (id) => {
         onEdit(id);
     };
     const handleDelete = (id) => {
         onDelete(id);
     };
+    if (!author && !category) return "loading...";
     return (
-        <>
-            <tr>
-                <td scope="col">{article.name}</td>
-                <td scope="col">{article.author}</td>
-                <td scope="col">{article.category}</td>
-                <td scope="col">{new Date(article.date).toLocaleDateString()}</td>
-                <td scope="col">{article.rate}</td>
-                <td scope="col">
-                    <a
-                        role="button"
-                        className="text-secondary"
-                        onClick={() => handleEdit(article._id)}
-                    >
-                        Редактировать
-                    </a>
-                </td>
-                <td scope="col">
-                    <button
-                        className="btn btn-danger"
-                        onClick={() => handleDelete(article._id)}
-                    >Удалить</button>
-                </td>
-            </tr>
-        </>
+        <tr className="text-center">
+            <td scope="col"><b>{article.name}</b></td>
+            <td scope="col">{`${firstName} ${lastName}`}</td>
+            <td scope="col">{category.name}</td>
+            <td scope="col">{new Date(article.date).toLocaleDateString()}</td>
+            <td scope="col">{article.rate}</td>
+            <td scope="col">
+                <a
+                    role="button"
+                    className="text-secondary"
+                    onClick={() => handleEdit(article._id)}
+                >
+                    Редактировать
+                </a>
+            </td>
+            <td scope="col">
+                <button
+                    className="btn btn-danger"
+                    onClick={() => handleDelete(article._id)}
+                >Удалить</button>
+            </td>
+        </tr>
     );
 };
